🏗️ Refactor Flask application to modern modular architecture
## Overview
Complete refactor of monolithic 2419-line app.py into clean, testable, maintainable modules.

## 📁 New Architecture
- **app.py** (67 lines): Clean entry point with dependency injection
- **config.py**: Centralized configuration management
- **routes/**: API endpoints and WebSocket handlers
- **services/**: Business logic layer (metadata, history, TTS)
- **utils/**: Reusable utilities (security, file I/O, text processing)
- **tests/**: Comprehensive test suite (31 tests)

## 🚀 Key Improvements
- ✅ **Modular design** - separation of concerns
- ✅ **Security hardening** - path traversal prevention
- ✅ **Atomic file operations** - prevent data corruption
- ✅ **Error handling** - robust error recovery
- ✅ **Type hints** - better code documentation
- ✅ **Test coverage** - 31 passing unit tests
- ✅ **Performance** - reduced memory footprint and complexity

## 🔧 Technical Changes
- Eliminated duplicate constants and global variables
- Implemented proper dependency injection pattern
- Added comprehensive path validation for security
- Created atomic file write operations for reliability
- Extracted business logic into service layer
- Added WebSocket handler abstraction
- Implemented proper configuration management

## 🧪 Testing
- **Config tests**: Validation of configuration system
- **Utils tests**: Security, text processing, file I/O
- **Service tests**: Metadata, history, TTS functionality
- **Integration tests**: End-to-end application flow
- **31 tests passing** with comprehensive coverage

## 📊 Impact
- **Before**: 2419 lines, 100+ functions in one file
- **After**: 7 focused modules, clean separation
- **Maintainability**: 📈 Significantly improved
- **Testability**: 📈 Fully testable architecture
- **Security**: 📈 Hardened against common attacks

## 🔄 Backward Compatibility
- All existing API endpoints preserved
- WebSocket interface unchanged
- Legacy function wrappers for Liquidsoap integration
- Drop-in replacement for existing deployment
