+++ ui/services/__init__.py
@@ -1,0 +1,8 @@
+"""
+Service layer for AI Radio business logic
+"""
+from .metadata import MetadataService
+from .history import HistoryService
+from .tts import TTSService
+
+__all__ = ['MetadataService', 'HistoryService', 'TTSService']
