+++ ui/tests/__init__.py
@@ -1,0 +1,3 @@
+"""
+Test suite for AI Radio Flask application
+"""
