+++ ui/tests/run_tests.py
@@ -1,0 +1,68 @@
+#!/usr/bin/env python3
+"""
+Test runner for AI Radio Flask application
+"""
+import sys
+import unittest
+import os
+from pathlib import Path
+
+# Add parent directory to path for imports
+sys.path.insert(0, str(Path(__file__).parent.parent))
+
+def run_all_tests():
+    """Run all tests in the test suite"""
+    
+    # Discover and run all tests
+    loader = unittest.TestLoader()
+    start_dir = Path(__file__).parent
+    suite = loader.discover(start_dir, pattern='test_*.py')
+    
+    # Run tests with verbose output
+    runner = unittest.TextTestRunner(verbosity=2, stream=sys.stdout)
+    result = runner.run(suite)
+    
+    # Print summary
+    print("\n" + "="*70)
+    print("TEST SUMMARY")
+    print("="*70)
+    print(f"Tests run: {result.testsRun}")
+    print(f"Failures: {len(result.failures)}")
+    print(f"Errors: {len(result.errors)}")
+    print(f"Skipped: {len(result.skipped) if hasattr(result, 'skipped') else 0}")
+    
+    if result.failures:
+        print("\nFAILURES:")
+        for test, traceback in result.failures:
+            print(f"- {test}")
+    
+    if result.errors:
+        print("\nERRORS:")
+        for test, traceback in result.errors:
+            print(f"- {test}")
+    
+    success = len(result.failures) == 0 and len(result.errors) == 0
+    print(f"\nOverall result: {'PASS' if success else 'FAIL'}")
+    
+    return success
+
+def run_specific_test(test_module):
+    """Run tests from a specific module"""
+    loader = unittest.TestLoader()
+    suite = loader.loadTestsFromName(test_module)
+    
+    runner = unittest.TextTestRunner(verbosity=2)
+    result = runner.run(suite)
+    
+    return len(result.failures) == 0 and len(result.errors) == 0
+
+if __name__ == '__main__':
+    if len(sys.argv) > 1:
+        # Run specific test module
+        test_module = sys.argv[1]
+        success = run_specific_test(test_module)
+    else:
+        # Run all tests
+        success = run_all_tests()
+    
+    sys.exit(0 if success else 1)
